+++ src/Firebase/ServiceAccount.php
@@ -6,8 +6,6 @@ namespace Kreait\Firebase;
 
 /**
  * @internal
- *
- * @todo Add #[SensitiveParameter] attributes once the minimum required PHP version is >=8.2
  */
 final class ServiceAccount
 {
@@ -15,26 +13,37 @@ final class ServiceAccount
         /** @var non-empty-string */
         public string $type,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $projectId,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $clientEmail,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $clientId,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $privateKey,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $privateKeyId,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $authUri,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $tokenUri,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $authProviderX509CertUrl,
         /** @var non-empty-string */
+        #[\SensitiveParameter]
         public string $clientX509CertUrl,
         /** @var non-empty-string|null */
+        #[\SensitiveParameter]
         public ?string $quotaProjectId = null,
         /** @var non-empty-string|null */
+        #[\SensitiveParameter]
         public ?string $universeDomain = null,
     ) {
     }